# docs/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "FA26SE103 Operations API",
    "version": "v1"
  },
  "paths": {
...
    "/api/zones/{zoneId}/monitoring": {
      "get": {
        "tags": [
          "Monitoring"
        ],
        "parameters": [
          {
            "name": "zoneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringConfigurationView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringConfigurationView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringConfigurationView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Monitoring"
        ],
        "parameters": [
          {
            "name": "zoneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitoringRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitoringRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MonitoringRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringConfigurationView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringConfigurationView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringConfigurationView"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Monitoring"
        ],
        "parameters": [
          {
            "name": "zoneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitoringDeleteRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitoringDeleteRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MonitoringDeleteRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/zones/{zoneId}/monitoring/review": {
      "get": {
        "tags": [
          "Monitoring"
        ],
        "parameters": [
          {
            "name": "zoneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringReviewView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringReviewView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringReviewView"
                }
              }
            }
          }
        }
      }
    },
    "/api/zones/{zoneId}/monitoring/activate": {
      "post": {
        "tags": [
          "Monitoring"
        ],
        "parameters": [
          {
            "name": "zoneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitoringActivationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitoringActivationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MonitoringActivationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringConfigurationView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringConfigurationView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringConfigurationView"
                }
              }
            }
          }
        }
      }
    },
    "/api/zones/{zoneId}/monitoring/deactivate": {
      "post": {
        "tags": [
          "Monitoring"
        ],
        "parameters": [
          {
            "name": "zoneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitoringActivationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitoringActivationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MonitoringActivationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringConfigurationView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringConfigurationView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringConfigurationView"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MonitoringActivationRequest": {
        "type": "object",
        "properties": {
          "expectedUpdatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "MonitoringCameraView": {
        "type": "object",
        "properties": {
          "cameraId": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "mappingStatus": {
            "type": "string",
            "nullable": true
          },
          "roiPolygon": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Point"
            },
            "nullable": true
          },
          "sourceType": {
            "type": "string",
            "nullable": true
          },
          "protocol": {
            "type": "string",
            "nullable": true
          },
          "isEnabled": {
            "type": "boolean"
          },
          "lastTestResult": {
            "type": "string",
            "nullable": true
          },
          "lastTestedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ready": {
            "type": "boolean"
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MonitoringIssue"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MonitoringConfigurationView": {
        "type": "object",
        "properties": {
          "configId": {
            "type": "string",
            "format": "uuid"
          },
          "zoneId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "confidenceThreshold": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "createdByUserId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MonitoringRuleView"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MonitoringDeleteRequest": {
        "type": "object",
        "properties": {
          "configId": {
            "type": "string",
            "format": "uuid"
          },
          "expectedUpdatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "MonitoringIssue": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MonitoringRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "confidenceThreshold": {
            "type": "number",
            "format": "double"
          },
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MonitoringRuleRequest"
            },
            "minItems": 1,
            "description": "At least one incident rule is required when saving; an empty array returns RULES_REQUIRED.",
            "nullable": true
          },
          "expectedUpdatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MonitoringReviewView": {
        "type": "object",
        "properties": {
          "configuration": {
            "$ref": "#/components/schemas/MonitoringConfigurationView"
          },
          "zone": {
            "$ref": "#/components/schemas/ZoneView"
          },
          "cameras": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MonitoringCameraView"
            },
            "nullable": true
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MonitoringIssue"
            },
            "nullable": true
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "canActivate": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "MonitoringRuleRequest": {
        "type": "object",
        "properties": {
          "incidentTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "warningThreshold": {
            "type": "number",
            "format": "double"
          },
          "criticalThreshold": {
            "type": "number",
            "format": "double"
          },
          "thresholdUnit": {
            "type": "string",
            "nullable": true
          },
          "sustainSec": {
            "type": "integer",
            "format": "int32"
          },
          "cooldownSec": {
            "type": "integer",
            "format": "int32"
          },
          "enabled": {
            "type": "boolean"
          },
          "parametersJson": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MonitoringRuleView": {
        "type": "object",
        "properties": {
          "ruleId": {
            "type": "string",
            "format": "uuid"
          },
          "incidentTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "incidentCode": {
            "type": "string",
            "nullable": true
          },
          "incidentName": {
            "type": "string",
            "nullable": true
          },
          "warningThreshold": {
            "type": "number",
            "format": "double"
          },
          "criticalThreshold": {
            "type": "number",
            "format": "double"
          },
          "thresholdUnit": {
            "type": "string",
            "nullable": true
          },
          "sustainSec": {
            "type": "integer",
            "format": "int32"
          },
          "cooldownSec": {
            "type": "integer",
            "format": "int32"
          },
          "enabled": {
            "type": "boolean"
          },
          "parametersJson": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Point": {
        "type": "object",
        "properties": {
          "x": {
            "type": "number",
            "format": "double"
          },
          "y": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "ZoneView": {
        "type": "object",
        "properties": {
          "zoneId": {
            "type": "string",
            "format": "uuid"
          },
          "floorId": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "zoneType": {
            "type": "string",
            "nullable": true
          },
          "mapPolygon": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Point"
            },
            "nullable": true
          },
          "colorHex": {
            "type": "string",
            "nullable": true
          },
          "areaM2": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}